backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-storage==2.14.0
google-auth==2.27.0
Pillow
python-multipart
requests
